+++ src/ADSB/ADSBVehicle.h
@@ -14,21 +14,20 @@
 #include <QtCore/QLoggingCategory>
 #include <QtPositioning/QGeoCoordinate>
 
-Q_DECLARE_LOGGING_CATEGORY(ADSBVehicleManagerLog)
+Q_DECLARE_LOGGING_CATEGORY(ADSBVehicleLog)
 
 class ADSBVehicle : public QObject
 {
     Q_OBJECT
 
-public:
-    enum {
-        CallsignAvailable =     1 << 1,
-        LocationAvailable =     1 << 2,
-        AltitudeAvailable =     1 << 3,
-        HeadingAvailable =      1 << 4,
-        AlertAvailable =        1 << 5,
-    };
+    Q_PROPERTY(int              icaoAddress READ icaoAddress    CONSTANT)
+    Q_PROPERTY(QString          callsign    READ callsign       NOTIFY callsignChanged)
+    Q_PROPERTY(QGeoCoordinate   coordinate  READ coordinate     NOTIFY coordinateChanged)
+    Q_PROPERTY(double           altitude    READ altitude       NOTIFY altitudeChanged)     // NaN for not available
+    Q_PROPERTY(double           heading     READ heading        NOTIFY headingChanged)      // NaN for not available
+    Q_PROPERTY(bool             alert       READ alert          NOTIFY alertChanged)        // Collision path
 
+public:
     typedef struct {
         uint32_t        icaoAddress;    // Required
         QString         callsign;
@@ -39,14 +38,15 @@ public:
         uint32_t        availableFlags;
     } ADSBVehicleInfo_t;
 
-    ADSBVehicle(const ADSBVehicleInfo_t & vehicleInfo, QObject* parent);
+    ADSBVehicle(const ADSBVehicleInfo_t &vehicleInfo, QObject* parent);
 
-    Q_PROPERTY(int              icaoAddress READ icaoAddress    CONSTANT)
-    Q_PROPERTY(QString          callsign    READ callsign       NOTIFY callsignChanged)
-    Q_PROPERTY(QGeoCoordinate   coordinate  READ coordinate     NOTIFY coordinateChanged)
-    Q_PROPERTY(double           altitude    READ altitude       NOTIFY altitudeChanged)     // NaN for not available
-    Q_PROPERTY(double           heading     READ heading        NOTIFY headingChanged)      // NaN for not available
-    Q_PROPERTY(bool             alert       READ alert          NOTIFY alertChanged)        // Collision path
+    enum {
+        CallsignAvailable =     1 << 1,
+        LocationAvailable =     1 << 2,
+        AltitudeAvailable =     1 << 3,
+        HeadingAvailable =      1 << 4,
+        AlertAvailable =        1 << 5,
+    };
 
     int             icaoAddress (void) const { return static_cast<int>(_icaoAddress); }
     QString         callsign    (void) const { return _callsign; }
@@ -55,17 +55,17 @@ public:
     double          heading     (void) const { return _heading; }
     bool            alert       (void) const { return _alert; }
 
-    void update(const ADSBVehicleInfo_t & vehicleInfo);
+    void update(const ADSBVehicleInfo_t &vehicleInfo);
 
     /// check if the vehicle is expired and should be removed
     bool expired();
 
 signals:
-    void coordinateChanged  ();
-    void callsignChanged    ();
-    void altitudeChanged    ();
-    void headingChanged     ();
-    void alertChanged       ();
+    void coordinateChanged();
+    void callsignChanged();
+    void altitudeChanged();
+    void headingChanged();
+    void alertChanged();
 
 private:
     uint32_t        _icaoAddress;
@@ -81,4 +81,3 @@ private:
 };
 
 Q_DECLARE_METATYPE(ADSBVehicle::ADSBVehicleInfo_t)
-
